# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EdeQR LMS API",
        "description": "bridge digital LMS projesi için REST API dokümantasyonu",
        "contact": {
            "email": "ahmet@aciksari.net"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/schools/{school}/students/createuser": {
            "post": {
                "tags": [
                    "SchoolStudentProfile"
                ],
                "summary": "Yeni öğrenci oluştur (sadece User tablosuna kaydeder)",
                "operationId": "38fa6e7e82ac4af723605d588b497c31",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreStudentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Öğrenci başarıyla oluşturuldu"
                    },
                    "403": {
                        "description": "Yetkisiz işlem"
                    }
                }
            }
        },
        "/api/schools/{school}/students/{user}/completeprofile": {
            "post": {
                "tags": [
                    "SchoolStudentProfile"
                ],
                "summary": "Öğrencinin profil, veli ve sağlık bilgilerini tamamlar",
                "operationId": "ae748ae469dcb542d4e8740ff502d10f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompleteStudentProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profil başarıyla oluşturuldu"
                    }
                }
            }
        },
        "/api/schools/{school}/students/{id}": {
            "get": {
                "tags": [
                    "SchoolStudentProfile"
                ],
                "summary": "Öğrenci detayını getirir",
                "operationId": "cf7890c15c62df52fb92825b3cf92c12",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı"
                    }
                }
            },
            "delete": {
                "tags": [
                    "SchoolStudentProfile"
                ],
                "summary": "Öğrenciyi sistemden siler (User + ilişkili tüm profiller)",
                "operationId": "94257c6d8ea9e97645a00d7e59534189",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeleteStudentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Öğrenci başarıyla silindi"
                    },
                    "403": {
                        "description": "Yetkisiz işlem"
                    },
                    "404": {
                        "description": "Öğrenci bulunamadı"
                    }
                }
            }
        },
        "/api/schools/{school}/students/byclass/{classModel}": {
            "get": {
                "tags": [
                    "SchoolStudentProfile"
                ],
                "summary": "Belirli bir sınıfa ait öğrencileri listeler",
                "operationId": "1a5c9b77ddaef5fb13b891cd45551d4c",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "classModel",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı"
                    }
                }
            }
        },
        "/api/schools/{school}/students": {
            "get": {
                "tags": [
                    "SchoolStudentProfile"
                ],
                "summary": "Okuldaki tüm öğrencileri listeler",
                "operationId": "d2c9878f9f86a191ccc06723d43e118d",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı"
                    }
                }
            }
        },
        "/api/schools/{school}/students/{id}/update": {
            "put": {
                "tags": [
                    "SchoolStudentProfile"
                ],
                "summary": "Öğrenciyi günceller (hem user hem profile)",
                "operationId": "1b58dbe5f2214fceecb4e24b1b6b89a1",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateStudentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Öğrenci başarıyla güncellendi"
                    },
                    "403": {
                        "description": "Yetkisiz işlem"
                    },
                    "404": {
                        "description": "Öğrenci bulunamadı"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CompleteStudentProfileRequest": {
                "required": [
                    "schoolId",
                    "active_class_model_id",
                    "birth_date",
                    "student_number",
                    "tc_no"
                ],
                "properties": {
                    "schoolId": {
                        "type": "integer",
                        "example": 1
                    },
                    "active_class_model_id": {
                        "type": "integer",
                        "example": 12
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2012-05-10"
                    },
                    "student_number": {
                        "type": "string",
                        "example": "2025A12"
                    },
                    "tc_no": {
                        "type": "string",
                        "example": "12345678901"
                    },
                    "parent_name": {
                        "type": "string",
                        "example": "Ali Yılmaz"
                    },
                    "parent_phone": {
                        "type": "string",
                        "example": "05551112233"
                    },
                    "gender": {
                        "type": "string",
                        "example": "erkek"
                    },
                    "address": {
                        "type": "string",
                        "example": "Ankara, Türkiye"
                    }
                },
                "type": "object"
            },
            "DeleteStudentRequest": {
                "required": [
                    "confirm"
                ],
                "properties": {
                    "confirm": {
                        "description": "Silme işlemini onaylamak için true gönderilmelidir.",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "StoreStudentRequest": {
                "required": [
                    "name",
                    "userName",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Ahmet Yılmaz"
                    },
                    "userName": {
                        "type": "string",
                        "example": "ahmetyilmaz"
                    },
                    "email": {
                        "type": "string",
                        "example": "ahmet@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
            "UpdateStudentRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Ahmet Can"
                    },
                    "email": {
                        "type": "string",
                        "example": "ahmetcan@example.com"
                    },
                    "userName": {
                        "type": "string",
                        "example": "ahmetcan123"
                    },
                    "password": {
                        "type": "string",
                        "example": "YeniŞifre123"
                    },
                    "phone": {
                        "type": "string",
                        "example": "05554443322"
                    },
                    "address": {
                        "type": "string",
                        "example": "İstanbul"
                    },
                    "parent_name": {
                        "type": "string",
                        "example": "Ali Can"
                    },
                    "parent_phone": {
                        "type": "string",
                        "example": "05551112233"
                    },
                    "blood_type": {
                        "type": "string",
                        "example": "A+"
                    },
                    "health_insurance": {
                        "type": "string",
                        "example": "SGK"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT veya Sanctum tokenini \"Bearer {token}\" formatında girin"
            },
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "1": {
            "bearerAuth": []
        }
    }
}